refactor: move home's doamin to domain module
