+++ feat/home/src/main/java/com/emotionstorage/home/data/repoImpl/HomeRepositoryImpl.kt
@@ -3,8 +3,8 @@ package com.emotionstorage.home.data.repoImpl
 import com.emotionstorage.domain.common.DataState
 import com.emotionstorage.home.data.dataSource.HomeRemoteDataSource
 import com.emotionstorage.home.data.modelMapper.HomeMapper
-import com.emotionstorage.home.domain.model.Home
-import com.emotionstorage.home.domain.repo.HomeRepository
+import com.emotionstorage.domain.model.Home
+import com.emotionstorage.domain.repo.HomeRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
